Implement support for abstract memory queues
Enable with LAVATUBE_VIRTUAL_QUEUES=1

This option presents the application with a virtual queue
interface with one queue family containing two queues, no
matter what the host system has. On replay, if the used
queue family is stored as virtual, then the queue used is
remapped to the first found graphics queue and any excess
queues are remapped to its first queue.