Add architecture for delayed writes
A tiny bit of time travel is now possible while streaming
bits to disk. The new write_later_*() functions freezes
processing of data and gives you a pointer for writing to
a later point in the data stream. The thaw() function now
unfreezes the data stream and invalidates all previous
pointers obtained this way.