Memory map uncompressed data instead of using a chunk list
Before this commit we had uncompressed data stored in a
list of chunks, access protected by mutexes. Now we store
this data in an anonymous memory map the size of the full
uncompressed file size, with a readahead thread filling it
up hopefully before we need it. Once a command has been
fully processed, we tell the OS to free any memory behind
the current position. If we catch up to the readahead
thread, we stall until it has had time to give us more
data.

The memory freeing will be in its own commit to make
bisecting of memory-related issue easier.

This design also allows a very nice implementation of
preloading. Once we get to the measurement framerange, we
wait for the readahead buffer to fill up, then we change
the stall on catch up into an abort. The user then has to
modify the size of the readahead buffer so that we no
longer stall during the frame range, or use an option for
replay to allow stalls. This is also yet to be implemented.

It is a lockless design, and most of the mutex usage for
file reading is now gone. This gives up to 50% reduction in
CPU usage for some workloads.

And the code is much smaller and cleaner!