Increase sandbox security
We now remove write access to the filesystem on sandbox
level three, except to current working directory and any
directories underneath it. We also remove network access
on level two. In order to make this work, we need to
initialize the window system connection before we enter
level two of the sandbox - and keep it open until we are
all done.