Finally wire up the explicit host updates support
This is similar to gfxr's "assisted" memory tracking, where
the capture layer relies on the application to explicitly
flush host memory before it is used on the GPU device.

LAVATUBE_TRUST_HOST_FLUSHING can be set to active it, or the
experimental VK_ARM_explicit_host_updates extension can be
used to enable it.